rename prefillIssueWithTemplate to autoSelectionBetweenBrackets
