+++ src/plugin/autoSelectionBetweenBrackets.js
